Rename QuasiNonsingularOrbitalElements
