+++ s2e-ff/test/TestQuasiNonsingularOrbitalElements.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "../src/Library/Orbit/QuasiNonsingularOrbitalElements.hpp"
+#include "../src/Library/Orbit/quasi_nonsingular_orbital_elements.hpp"
 
 TEST(QuasiNonsingularOrbitalElements, DefaultConstructor) {
   QuasiNonsingularOrbitalElements qn_oe;
